Backend/Capstone/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
websockets==12.0
protobuf==5.27.2
python-dotenv==1.0.1

